requirements.txt
coverage==4.4.1
Django==1.11.5
django-cors-headers==2.1.0
django-extensions==1.9.1
django-nose==1.4.5
djangorestframework==3.6.4
nose==1.3.7
psycopg2==2.7.3.1
pytz==2017.2
six==1.10.0
